Re arrange the header move to any root files

+++ Site/admin/delete-tag-proc.php
@@ -1,5 +1,13 @@
-<?php include("include/header.php");?>
-<?php
+<?php 
+include ("include/header.php");
+@session_start();
+$rootpath = "../";
+include ($rootpath . "lib/db.php");
+include ($rootpath . "lib/conn.inc.php");
+include ($rootpath . "lib/func_date.php");
+if (!$db -> open()) {
+	die($db -> error());
+}
 $tagId = $_POST['tagId'];
 
 $strSQL="DELETE FROM TAG WHERE ID='$tagId'";
